+++ packages/type-storage/src/fromMetadata/getHasher.ts
@@ -6,36 +6,15 @@ import { Text } from '@polkadot/types';
 import { StorageHasher } from '@polkadot/types/primitive';
 import { blake2AsU8a, xxhashAsU8a } from '@polkadot/util-crypto';
 
-type HasherInput = string | Buffer | Uint8Array;
+type HasherInput = Buffer | Uint8Array;
 
 export type HasherFunction = (data: HasherInput) => Uint8Array;
 
-function toStorageHasher (text: Text): StorageHasher {
-  switch (text.toString()) {
-    case 'blake2_128':
-      return new StorageHasher('Blake2_128');
-    case 'blake2_256':
-      return new StorageHasher('Blake2_256');
-    case 'twox_128':
-      return new StorageHasher('Twox128');
-    case 'twox_256':
-      return new StorageHasher('Twox256');
-    case 'twox_64_concat':
-      return new StorageHasher('Twox64Concat');
-    default:
-      throw new Error(`Invalid Storage hasher: ${text.toString()}`);
-  }
-}
-
-export default function getHasher (hasherInput?: StorageHasher | Text): HasherFunction {
+export default function getHasher (hasher?: StorageHasher): HasherFunction {
   // This one is the default for PlainType storage keys
-  if (!hasherInput) {
+  if (!hasher) {
     return (data: HasherInput) => xxhashAsU8a(data, 128);
   }
-
-  const hasher: StorageHasher = hasherInput instanceof Text
-    ? toStorageHasher(hasherInput)
-    : hasherInput;
 
   if (hasher.isBlake2128) {
     return (data: HasherInput) => blake2AsU8a(data, 128);
