+++ packages/types/src/Metadata/v4/toV5.ts
@@ -3,21 +3,36 @@
 // of the Apache-2.0 license. See the LICENSE file for details.
 
 import { Option, Vector } from '../../codec';
+import Text from '../../primitive/Text';
 import MetadataV4 from './Metadata';
 import StorageHasher from '../../primitive/StorageHasher';
 import MetadataV5 from '../v5';
-import { StorageFunctionMetadata as StorageFunctionMetadataV3 } from '../v3/Storage';
-import { DoubleMapType, MapType, StorageFunctionMetadata, StorageFunctionType } from '../v4/Storage';
+import { StorageFunctionMetadata as StorageFunctionMetadataV4 } from '../v4/Storage';
+import { DoubleMapType, MapType, StorageFunctionMetadata, StorageFunctionType } from '../v5/Storage';
 
 /**
  * Convert V4 StorageFunction to V5 StorageFunction
  */
-function toV5StorageFunction (storageFn: StorageFunctionMetadataV3): StorageFunctionMetadata {
+function toV5StorageFunction (storageFn: StorageFunctionMetadataV4): StorageFunctionMetadata {
   const { documentation, fallback, modifier, name, type } = storageFn;
 
-  // Convert the old type to the new type: there is one new field
-  // called `hasher`, which we initialize to xxHash (the default in
-  // v3).
+  function toStorageHasher (text: Text): StorageHasher {
+    switch (text.toString()) {
+      case 'blake2_128':
+        return new StorageHasher('Blake2_128');
+      case 'blake2_256':
+        return new StorageHasher('Blake2_256');
+      case 'twox_128':
+        return new StorageHasher('Twox128');
+      case 'twox_256':
+        return new StorageHasher('Twox256');
+      case 'twox_64_concat':
+        return new StorageHasher('Twox64Concat');
+      default:
+        throw new Error(`Invalid Storage hasher: ${text.toString()}`);
+    }
+  }
+
   const [newType, index] = type.isPlainType
     ? [type, 0]
     : type.isMap
@@ -32,7 +47,7 @@ function toV5StorageFunction (storageFn: StorageFunctionMetadataV3): StorageFunc
         key1: type.asDoubleMap.key1,
         key2: type.asDoubleMap.key2,
         value: type.asDoubleMap.value,
-        key2Hasher: type.asDoubleMap.key2Hasher
+        key2Hasher: toStorageHasher(type.asDoubleMap.key2Hasher)
       }), 2];
 
   return new StorageFunctionMetadata({
